+++ sema4ai/tests/sema4ai_code_tests/test_compute_data_sources_state/package/data_sources.py
@@ -2,7 +2,7 @@ from typing import Annotated
 
 from sema4ai.data import DataSource, DataSourceSpec
 
-# Generated data source (can be configured automatically)
+# Generated data source (files:can be configured automatically)
 CustomersDataSource = Annotated[
     DataSource,
     DataSourceSpec(
@@ -10,6 +10,23 @@ CustomersDataSource = Annotated[
         file="files/customers.csv",
         engine="files",
         description="Data source for customers.",
+    ),
+]
+
+# Generated data source (models:can be configured automatically)
+PredictDataSource = Annotated[
+    DataSource,
+    DataSourceSpec(
+        model_name="predict_compute_data_sources_state",
+        engine="prediction:lightwood",
+        description="Predict something.",
+        sql="""
+CREATE MODEL models.predict_compute_data_sources_state
+(SELECT * FROM files.customers)
+PREDICT country
+ORDER BY first_name
+WINDOW 8
+HORIZON 4;""",
     ),
 ]